Modified the model, moved parameters to top of the file, and fixed latest prediction.
I reduced the amount of items we use to predict. Removing items like vial of water may help reduce noise.

I moved BATCH_SIZE to the top of the file to make it easier to change. Eventually, these may be made into arguments.

I changed the end of the program which makes a prediction on the latest data. Also, it gives the RMSE of testing data before making a prediction.

The program also now provides an overview of testing target data (looking at the standard deviation may help understand the model better)