I improved the model structure using resources from the internet (more to be researched and learned on).
I improved the end-of-training workflow by adding a comparison between std and rmse. I also removed the debug prints at the end.

I reactivated the python plotting lines to see how the model is predicting vs the actual values. The model is doing fairly well, but it is hard to see if it is getting the movement correct (at least getting the + or - correct).

More testing remains to be done on choosing a good batch size and epochs.